+++ docs/references/orchestration/prefect/prefect_flows.md
@@ -1,0 +1,19 @@
+::: viadot.orchestration.prefect.flows.cloud_for_customers_to_adls
+
+::: viadot.orchestration.prefect.flows.cloud_for_customers_to_databricks
+
+::: viadot.orchestration.prefect.flows.exchange_rates_to_adls
+
+::: viadot.orchestration.prefect.flows.sap_to_redshift_spectrum
+
+::: viadot.orchestration.prefect.flows.sharepoint_to_adls
+
+::: viadot.orchestration.prefect.flows.sharepoint_to_databricks
+
+::: viadot.orchestration.prefect.flows.sharepoint_to_redshift_spectrum
+
+::: viadot.orchestration.prefect.flows.sharepoint_to_s3
+
+::: viadot.orchestration.prefect.flows.transform
+
+::: viadot.orchestration.prefect.flows.transform_and_catalog
