+++ docs/references/orchestration/prefect/prefect_tasks.md
@@ -1,0 +1,30 @@
+
+::: viadot.orchestration.prefect.tasks.adls_upload
+
+::: viadot.orchestration.prefect.tasks.df_to_adls
+
+::: viadot.orchestration.prefect.tasks.cloud_for_customers_to_df
+
+::: viadot.orchestration.prefect.tasks.df_to_databricks
+
+::: viadot.orchestration.prefect.tasks.dbt_task
+
+::: viadot.orchestration.prefect.tasks.exchange_rates_to_df
+
+::: viadot.orchestration.prefect.tasks.clone_repo
+
+::: viadot.orchestration.prefect.tasks.luma_ingest_task
+
+::: viadot.orchestration.prefect.tasks.df_to_redshift_spectrum
+
+::: viadot.orchestration.prefect.tasks.s3_upload_file
+
+::: viadot.orchestration.prefect.tasks.get_endpoint_type_from_url
+
+::: viadot.orchestration.prefect.tasks.scan_sharepoint_folder
+
+::: viadot.orchestration.prefect.tasks.sharepoint_download_file
+
+::: viadot.orchestration.prefect.tasks.sharepoint_to_df
+
+::: viadot.orchestration.prefect.tasks.validate_and_reorder_dfs_columns
