+++ src/viadot/orchestration/prefect/flows/__init__.py
@@ -1,8 +1,8 @@
 """Import flows."""
 
-from .cloud_for_customers import (  # noqa: F401
-    cloud_for_customers_to_adls,
-    cloud_for_customers_to_databricks,
+from .cloud_for_customers_to_adls import cloud_for_customers_to_adls  # noqa: F401
+from .cloud_for_customers_to_databricks import (
+    cloud_for_customers_to_databricks,  # noqa: F401
 )
 from .exchange_rates_to_adls import exchange_rates_to_adls  # noqa: F401
 from .exchange_rates_to_databricks import exchange_rates_to_databricks  # noqa: F401
